@@ -11,7 +11,7 @@
     "test": "jest",
     "coverage": "cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
     "lint": "eslint cli/ src/ config/",
-    "lint:fix": "eslint --fix cli/ src/ config/",
+    "lint:fix": "npm run lint -- --fix",
     "watch:test": "npm test -- --watch --coverage=false"
   },
   "author": "Isaac Suttell <isaac.suttell@sony.com>",
@@ -71,11 +71,14 @@
   "devDependencies": {
     "babel-eslint": "^7.2.3",
     "coveralls": "^2.13.1",
-    "eslint": "^4.3.0",
-    "eslint-config-ship-components-base": "^1.1.0",
-    "eslint-plugin-react": "^7.12.3",
+    "eslint": "^5.3.0",
+    "eslint-config-airbnb-base": "^13.2.0",
+    "eslint-config-prettier": "^6.0.0",
+    "eslint-plugin-import": "^2.18.0",
+    "eslint-plugin-prettier": "^3.1.0",
     "jest": "^21.2.1",
-    "jest-junit": "^3.1.0"
+    "jest-junit": "^3.1.0",
+    "prettier": "^1.18.2"
   },
   "jest": {
     "collectCoverage": true,
