@@ -1,39 +1,40 @@
 #!/usr/bin/env node
-/*******************************************************************************
+/** *****************************************************************************
  * CLI
  *
  * @author       Isaac Suttell <isaac_suttell@playstation.sony.com>
  * @file         Command Line Interface for the server
- ******************************************************************************/
+ ***************************************************************************** */
 
-var program = require('commander');
-var pkg = require(require('path').resolve(__dirname + '/../package.json'));
-var Create = require('./lib/Create');
-var path = require('path');
+const program = require('commander');
+const pkg = require(require('path').resolve(`${__dirname}/../package.json`));
+const path = require('path');
+const Create = require('./lib/Create');
 
-program.version(pkg.version)
+program
+  .version(pkg.version)
   .option('-v, --verbose', 'Display some extra details')
   .usage('[options] <name>')
-  .action(function(name, config){
+  .action((name, config) => {
     if (typeof name !== 'string') {
       program.outputHelp();
       process.exit(0);
     }
 
-    var filename = path.resolve(process.cwd(), './server/controllers/', name + '.js');
+    const filename = path.resolve(process.cwd(), './server/controllers/', `${name}.js`);
 
     if (config.verbose) {
       console.log('[VERBOSE] Saving to %s', filename);
     }
 
     Create.controller(filename, name)
-      .then(function(){
+      .then(() => {
         console.log('');
         console.log('%s successfully created', name);
         console.log('Do not forget to add it to your list of controllers in config/default.js');
         console.log('');
       })
-      .catch(function(err){
+      .catch(err => {
         console.error(err.stack);
         process.exit(1);
       });
