@@ -1,40 +1,40 @@
 #!/usr/bin/env node
-/*******************************************************************************
+/** *****************************************************************************
  * CLI
  *
  * @author       Isaac Suttell <isaac_suttell@playstation.sony.com>
  * @file         Command Line Interface for the server
- ******************************************************************************/
+ ***************************************************************************** */
 
-var program = require('commander');
-var pkg = require(require('path').resolve(__dirname + '/../package.json'));
-var Create = require('./lib/Create');
-var path = require('path');
-var Promise = require('bluebird');
-var mkdirp = require('mkdirp');
-var fs = require('fs');
-var spawn = require('child_process').spawn;
+const program = require('commander');
+const pkg = require(require('path').resolve(`${__dirname}/../package.json`));
+const path = require('path');
+const Promise = require('bluebird');
+const mkdirp = require('mkdirp');
+const fs = require('fs');
+const { spawn } = require('child_process');
+const Create = require('./lib/Create');
 
-var SERVER_ROOT = path.resolve(process.cwd(), './server');
+const SERVER_ROOT = path.resolve(process.cwd(), './server');
 
 function runCmd(cmd, args, config) {
-  return new Promise(function(resolve, reject){
-    var command = spawn(cmd, args);
+  return new Promise((resolve, reject) => {
+    const command = spawn(cmd, args);
 
-    command.stdout.on('data', function(data){
+    command.stdout.on('data', data => {
       if (config.verbose) {
         console.log(data.toString());
       }
     });
 
-    command.stderr.on('data', function(data){
+    command.stderr.on('data', data => {
       if (config.verbose) {
         console.log(data.toString());
       }
     });
 
-    command.on('close', function(code){
-      if(code === 0) {
+    command.on('close', code => {
+      if (code === 0) {
         resolve();
       } else {
         reject(new Error('Non zero exit code'));
@@ -44,9 +44,9 @@ function runCmd(cmd, args, config) {
 }
 
 function createDirectory(dir) {
-  return new Promise(function(resolve, reject){
-    mkdirp(dir, function(err){
-      if(err) {
+  return new Promise((resolve, reject) => {
+    mkdirp(dir, err => {
+      if (err) {
         reject(err);
       } else {
         resolve();
@@ -55,9 +55,9 @@ function createDirectory(dir) {
   });
 }
 
-function createDirectories(paths, config){
-  return Promise.each(paths, function(dir){
-    if(config.verbose) {
+function createDirectories(paths, config) {
+  return Promise.each(paths, dir => {
+    if (config.verbose) {
       console.log('Creating directory %s', dir);
     }
     return createDirectory(dir);
@@ -71,9 +71,9 @@ function createDirectories(paths, config){
  * @return {Promise}
  */
 function copyFile(original, target) {
-  return new Promise(function(resolve, reject){
-    var source = fs.createReadStream(original);
-    var dest = fs.createWriteStream(target);
+  return new Promise((resolve, reject) => {
+    const source = fs.createReadStream(original);
+    const dest = fs.createWriteStream(target);
 
     // Copy
     source.pipe(dest);
@@ -86,95 +86,95 @@ function copyFile(original, target) {
   });
 }
 
-program.version(pkg.version)
+program
+  .version(pkg.version)
   .option('-v, --verbose', 'Display some extra details')
   .usage('[options] <name>')
-  .action(function(name, config){
+  .action((name, config) => {
     if (typeof name !== 'string' || name.length === 0) {
       program.outputHelp();
       process.exit(0);
     }
 
-    var paths = [
+    const paths = [
       SERVER_ROOT,
       path.resolve(SERVER_ROOT, './controllers'),
       path.resolve(SERVER_ROOT, './middleware'),
       path.resolve(SERVER_ROOT, './models'),
       path.resolve(SERVER_ROOT, './lib'),
       path.resolve(SERVER_ROOT, './views'),
-      path.resolve(process.cwd(), './config')
+      path.resolve(process.cwd(), './config'),
     ];
 
     createDirectories(paths, config)
-      .then(function(){
-        var filename = path.resolve(process.cwd(), './server/controllers/IndexController.js');
+      .then(() => {
+        const filename = path.resolve(process.cwd(), './server/controllers/IndexController.js');
 
         if (config.verbose) {
           console.log('Creating controller %s', filename);
         }
 
-        return Create.controller(filename, 'IndexController').catch(function(err){
+        return Create.controller(filename, 'IndexController').catch(err => {
           console.log('There was a problem writing %s, skipping...', filename);
           if (config.verbose) {
             console.error(err);
           }
         });
       })
-      .then(function(){
-        var files = [
+      .then(() => {
+        const files = [
           {
             from: path.resolve(__dirname, '../config/default.js'),
-            to: path.resolve(process.cwd(), './config/default.js')
+            to: path.resolve(process.cwd(), './config/default.js'),
           },
           {
             from: path.resolve(__dirname, './templates/ceres.js'),
-            to: path.resolve(process.cwd(), './' + name + '.js')
+            to: path.resolve(process.cwd(), `./${name}.js`),
           },
           {
             from: path.resolve(__dirname, './templates/ceres-run.js'),
-            to: path.resolve(process.cwd(), './' + name + '-run.js')
+            to: path.resolve(process.cwd(), `./${name}-run.js`),
           },
           {
             from: path.resolve(__dirname, './templates/ceres-init.js'),
-            to: path.resolve(process.cwd(), './' + name + '-init.js')
-          }
+            to: path.resolve(process.cwd(), `./${name}-init.js`),
+          },
         ];
 
-        return Promise.each(files, function(file){
+        return Promise.each(files, file => {
           if (config.verbose) {
             console.log('Saving file to %s', file.to);
           }
           return copyFile(file.from, file.to);
         });
       })
-      .then(function(){
-
-        var commands = [
+      .then(() => {
+        const commands = [
           {
             cmd: 'npm',
-            args: ['init', '-y']
+            args: ['init', '-y'],
           },
           {
             cmd: 'npm',
-            args: ['install', 'ceres-framework@latest', '--save']
+            args: ['install', 'ceres-framework@latest', '--save'],
           },
           {
             cmd: 'chmod',
-            args: ['u+x', path.resolve(process.cwd(), './' + name + '.js')]
-          }
+            args: ['u+x', path.resolve(process.cwd(), `./${name}.js`)],
+          },
         ];
 
-        return Promise.each(commands, function(item){
+        return Promise.each(commands, item => {
           if (config.verbose) {
             console.log('Running %s %s', item.cmd, item.args.join(' '));
           }
           return runCmd(item.cmd, item.args, config);
         });
       })
-      .then(function(){
+      .then(() => {
         console.log('Success');
       })
-      .catch(function(err){
+      .catch(err => {
         console.error(err.stack);
         process.exit(1);
       });
