@@ -1,16 +1,18 @@
-var Promise = require('bluebird');
-var fs = require('fs');
-var ejs = require('ejs');
-var path = require('path');
+const Promise = require('bluebird');
+const fs = require('fs');
+const ejs = require('ejs');
+const path = require('path');
 
 /**
  * EJS Templates
  * @type {Object}
  */
-var Template = {
-  controller: fs.readFileSync(path.resolve(__dirname, '../templates/controller.ejs'), {encoding: 'utf8'}),
+const Template = {
+  controller: fs.readFileSync(path.resolve(__dirname, '../templates/controller.ejs'), {
+    encoding: 'utf8',
+  }),
 
-  model: fs.readFileSync(path.resolve(__dirname, '../templates/model.ejs'), {encoding: 'utf8'})
+  model: fs.readFileSync(path.resolve(__dirname, '../templates/model.ejs'), { encoding: 'utf8' }),
 };
 
 /**
@@ -19,35 +21,35 @@ var Template = {
  * @return {Promise}
  */
 function checkIfExists(filename) {
-  return new Promise(function(resolve, reject){
+  return new Promise((resolve, reject) => {
     // Check to see if we can access the file
-    fs.access(filename, function(err){
+    fs.access(filename, err => {
       if (err) {
         resolve();
       } else {
-        reject(new Error('File exists: ' + filename));
+        reject(new Error(`File exists: ${filename}`));
       }
     });
   });
 }
 
 module.exports = {
-
-  controller: function(filename, name){
-    return new Promise(function(resolve, reject){
-      var str = '';
+  controller(filename, name) {
+    return new Promise((resolve, reject) => {
+      let str = '';
       try {
         str = ejs.render(Template.controller, {
-          name: name
+          name,
         });
       } catch (err) {
-        return reject(err);
+        reject(err);
+        return undefined;
       }
 
       checkIfExists(filename)
-        .then(function(){
-          fs.writeFile(filename, str, function(err){
-            if(err) {
+        .then(() => {
+          fs.writeFile(filename, str, err => {
+            if (err) {
               reject(err);
             } else {
               resolve();
@@ -55,24 +57,26 @@ module.exports = {
           });
         })
         .catch(reject);
+      return undefined;
     });
   },
 
-  model: function(filename, name){
-    return new Promise(function(resolve, reject){
-      var str = '';
+  model(filename, name) {
+    return new Promise((resolve, reject) => {
+      let str = '';
       try {
         str = ejs.render(Template.model, {
-          name: name
+          name,
         });
       } catch (err) {
-        return reject(err);
+        reject(err);
+        return undefined;
       }
 
       checkIfExists(filename)
-        .then(function(){
-          fs.writeFile(filename, str, function(err){
-            if(err) {
+        .then(() => {
+          fs.writeFile(filename, str, err => {
+            if (err) {
               reject(err);
             } else {
               resolve();
@@ -80,6 +84,7 @@ module.exports = {
           });
         })
         .catch(reject);
+      return undefined;
     });
-  }
+  },
 };
