@@ -1,100 +1,96 @@
-'use strict';
-var wrapRoute = require('../../src/controllers/wrapRoute');
-var Promise = require('bluebird');
+const Promise = require('bluebird');
+const wrapRoute = require('../../src/controllers/wrapRoute');
 
-describe('wrapRoute', function(){
-  var ceres;
+describe('wrapRoute', () => {
+  let ceres;
 
-  beforeEach(function(){
+  beforeEach(() => {
     ceres = {
       log: {
-        _ceres: {
-          debug: function(){}
-        }
-      }
+        internal: {
+          debug() {},
+        },
+      },
     };
   });
 
-  it('should return a function', function(){
-    var handler = function() {};
-    var ctx = {};
-    var result;
-    expect(function(){
+  it('should return a function', () => {
+    function handler() {}
+    const ctx = {};
+    let result;
+    expect(() => {
       result = wrapRoute(handler, ctx, ceres);
     }).not.toThrow();
     expect(typeof result).toBe('function');
   });
 
-  it('should extend "this" with any custom properties', function(){
-    var result;
-    var handler = function() {
+  it('should extend "this" with any custom properties', () => {
+    let result;
+    function handler() {
       result = this;
+    }
+    const ctx = {
+      extend: true,
     };
-    var ctx = {
-      extend: true
-    };
-    expect(function(){
-      var wrappedFunction = wrapRoute(handler, ctx, ceres);
-      wrappedFunction()
-        .finally(() => {
-          expect(result.extend).toBe(ctx.extend);
-        });
+    expect(() => {
+      const wrappedFunction = wrapRoute(handler, ctx, ceres);
+      wrappedFunction().finally(() => {
+        expect(result.extend).toBe(ctx.extend);
+      });
     }).not.toThrow();
   });
 
-  it('should assign "req", "res", "next" to "this" of the handler', function(){
-    var result;
-    var handler = function() {
+  it('should assign "req", "res", "next" to "this" of the handler', () => {
+    let result;
+    function handler() {
       result = this;
-    };
-    var ctx = {};
+    }
+    const ctx = {};
 
-    var req = {};
-    var res = {};
-    var next = function(){};
-    expect(function(){
-      var wrappedFunction = wrapRoute(handler, ctx, ceres);
-      wrappedFunction(req, res, next)
-        .finally(() => {
-          expect(result.req).toBe(req);
-          expect(result.res).toBe(res);
-          expect(result.next).toBe(next);
-        });
+    const req = {};
+    const res = {};
+    const next = () => {};
+    expect(() => {
+      const wrappedFunction = wrapRoute(handler, ctx, ceres);
+      wrappedFunction(req, res, next).finally(() => {
+        expect(result.req).toBe(req);
+        expect(result.res).toBe(res);
+        expect(result.next).toBe(next);
+      });
     }).not.toThrow();
   });
 
-  it('should call the "next" handler if theres an error', function(){
-    var err = new Error('ERROR');
-    var handler = function() {
+  it('should call the "next" handler if theres an error', () => {
+    const err = new Error('ERROR');
+    const handler = () => {
       throw err;
     };
-    var ctx = {};
-    var next = jest.fn();
-    var wrappedFunction = wrapRoute(handler, ctx, ceres);
-    wrappedFunction({}, {}, next)
-      .then(() => {
-        expect(next).toHaveBeenCalledWith(err);
-      });
+    const ctx = {};
+    const next = jest.fn();
+    const wrappedFunction = wrapRoute(handler, ctx, ceres);
+    wrappedFunction({}, {}, next).then(() => {
+      expect(next).toHaveBeenCalledWith(err);
+    });
   });
 
-  it('should automatically "send" the result of a promise', function(done){
-    var expected = {
-      results: []
+  it('should automatically "send" the result of a promise', done => {
+    const expected = {
+      results: [],
     };
-    var handler = function(){
-      return new Promise(function(resolve){
+    const handler = () => {
+      return new Promise(resolve => {
         resolve(expected);
       });
     };
-    var ctx = {
-      send: jest.fn()
+    const ctx = {
+      send: jest.fn(),
     };
-    var res = {
-      writable: true
+    const res = {
+      writable: true,
     };
-    var fn = wrapRoute(handler, ctx, ceres);
-    fn({}, res, function(){})
-      .then(function(){
+    const fn = wrapRoute(handler, ctx, ceres);
+    fn({}, res, () => {})
+      .then(() => {
         expect(ctx.send).toHaveBeenCalledWith(expected);
       })
       .finally(() => {
@@ -102,26 +98,24 @@ describe('wrapRoute', function(){
       });
   });
 
-  it('should not write the response if is not writable', function(done){
-    var expected = {
-      results: []
+  it('should not write the response if is not writable', () => {
+    const expected = {
+      results: [],
     };
-    var handler = function(){
-      return new Promise(function(resolve){
+    function handler() {
+      return new Promise(resolve => {
         resolve(expected);
       });
+    }
+    const ctx = {
+      send: jest.fn(),
     };
-    var ctx = {
-      send: jest.fn()
+    const res = {
+      writable: false,
     };
-    var res = {
-      writable: false
-    };
-    var fn = wrapRoute(handler, ctx, ceres);
-    fn({}, res, function(){})
-      .then(function(){
-        expect(ctx.send).not.toHaveBeenCalled();
-        done();
-      });
+    const fn = wrapRoute(handler, ctx, ceres);
+    return fn({}, res, () => {}).then(() => {
+      expect(ctx.send).not.toHaveBeenCalled();
+    });
   });
 });
