@@ -2,6 +2,4 @@
  * Mock config for testing
  * @type    {Object}
  */
-module.exports = {
-
-};
+module.exports = {};
