@@ -2,8 +2,8 @@
  * Mock config for testing errors
  * @type    {Object}
  */
-module.exports = function(){
+module.exports = () => {
   return {
-    key: require('fs').readFileSync('does-not-exist')
+    key: require('fs').readFileSync('does-not-exist'),
   };
 };
