@@ -1,32 +1,30 @@
-'use strict';
+const merge = require('../../src/lib/merge');
 
-var merge = require('../../src/lib/merge');
-
-describe('merge', function(){
-  it('should export a function', function() {
+describe('merge', () => {
+  it('should export a function', () => {
     expect(typeof merge).toBe('function');
   });
 
-  it('should merge two objects', function() {
-    var obj1 = {
+  it('should merge two objects', () => {
+    const obj1 = {
       name: 'test',
-      test: false
+      test: false,
     };
 
-    var obj2 = {
-      test: true
+    const obj2 = {
+      test: true,
     };
 
-    var result = merge(obj1, obj2);
+    const result = merge(obj1, obj2);
 
     expect(result.test).toBe(obj2.test);
     expect(result.name).toBe(obj1.name);
   });
 
-  it('should modify the original object', function() {
-    var obj1 = {};
-    var obj2 = {
-      name: 'test'
+  it('should modify the original object', () => {
+    const obj1 = {};
+    const obj2 = {
+      name: 'test',
     };
 
     merge(obj1, obj2);
@@ -34,66 +32,66 @@ describe('merge', function(){
     expect(obj1.name).toBe(obj2.name);
   });
 
-  it('should return the result', function() {
-    var obj1 = {};
-    var obj2 = {
-      name: 'test'
+  it('should return the result', () => {
+    const obj1 = {};
+    const obj2 = {
+      name: 'test',
     };
 
-    var result = merge(obj1, obj2);
+    const result = merge(obj1, obj2);
 
     expect(result).toBe(obj1);
   });
 
-  it('should NOT deeply merge an array', function() {
-    var obj1 = {
-      arr: ['one']
+  it('should NOT deeply merge an array', () => {
+    const obj1 = {
+      arr: ['one'],
     };
-    var obj2 = {
-      arr: ['two', 'three']
+    const obj2 = {
+      arr: ['two', 'three'],
     };
 
-    var result = merge(obj1, obj2);
+    const result = merge(obj1, obj2);
 
     expect(result.arr.length).toBe(obj2.arr.length);
   });
 
-  it('should merge multiple objects', function() {
-    var obj1 = {
-      name: 'one'
+  it('should merge multiple objects', () => {
+    const obj1 = {
+      name: 'one',
     };
 
-    var obj2 = {
-      name: 'two'
+    const obj2 = {
+      name: 'two',
     };
 
-    var obj3 = {
-      name: 'three'
+    const obj3 = {
+      name: 'three',
     };
 
-    var result = merge(obj1, obj2, obj3);
+    const result = merge(obj1, obj2, obj3);
 
     expect(result.name).toBe(obj3.name);
   });
 
-  it('should ignored undefined values', function() {
-    var obj1 = {
+  it('should ignored undefined values', () => {
+    const obj1 = {
       port: 3000,
-      name: 'test'
+      name: 'test',
     };
 
-    var obj2 = {
-      port: void 0
+    const obj2 = {
+      port: void 0,
     };
 
-    var result = merge(obj1, obj2);
+    const result = merge(obj1, obj2);
 
     expect(result.port).toBe(obj1.port);
     expect(result.name).toBe(obj1.name);
   });
 
-  it('should deeply merge two objects', function() {
-    var obj1 = {
+  it('should deeply merge two objects', () => {
+    const obj1 = {
       name: 'test',
       test: false,
       config: {
@@ -101,27 +99,27 @@ describe('merge', function(){
         port: 3000,
         deepConfig: {
           db: 0,
-          ssl: false
+          ssl: false,
         },
         cache: {
-          type: 'test'
-        }
-      }
+          type: 'test',
+        },
+      },
     };
 
-    var obj2 = {
+    const obj2 = {
       test: true,
       config: {
         port: 5000,
         deepConfig: {
           db: 1,
-          ssl: true
+          ssl: true,
         },
-        cache: false
-      }
+        cache: false,
+      },
     };
 
-    var result = merge(obj1, obj2);
+    const result = merge(obj1, obj2);
 
     expect(result.config.port).toBe(obj2.config.port);
     expect(result.config.host).toBe(obj1.config.host);
