@@ -1,56 +1,53 @@
-'use strict';
+const Model = require('../../src/models/Model');
 
-var Model = require('../../src/models/Model');
+describe('Model', () => {
+  let ceres;
 
-describe('Model', function(){
-	var ceres;
+  beforeEach(() => {
+    ceres = {
+      config: {
+        db: {
+          type: 'bookshelf',
+        },
+      },
+      Database: {
+        bookshelf: {
+          Model: {
+            extend() {},
+          },
+          knex: {
+            raw() {},
+          },
+        },
+      },
+    };
+  });
 
-	beforeEach(function(){
-		ceres = {
-			config: {
-				db: {
-					type: 'bookshelf'
-				}
-			},
-			Database: {
-				bookshelf: {
-					Model: {
-						extend: function() {}
-					},
-					knex: {
-						raw: function() {}
-					}
-				}
-			}
-		};
-	});
+  it('should throw an error if it does not recognize the type', () => {
+    expect(() => {
+      new Model(ceres, {
+        type: 'badType',
+      });
+    }).toThrow();
+  });
 
-	it('should throw an error if it does not recognize the type', function(){
-		expect(function(){
-			new Model(ceres, {
-				type: 'badType'
-			});
-		}).toThrow();
-	});
+  it('should call the init function if provided', () => {
+    const spy = jest.fn();
+    const model = new Model(ceres, {
+      init: spy,
+    });
+    expect(model).toBeDefined();
+    expect(spy).toHaveBeenCalled();
+  });
 
-	it('should call the init function if provided', function(){
-		var spy = jest.fn();
-		var model = new Model(ceres, {
-			init: spy
-		});
-		expect(model).toBeDefined();
-		expect(spy).toHaveBeenCalled();
-	});
-
-	it('should have a extend function', function(){
-		var model;
-		expect(function(){
-			model = Model.extend.call(ceres, {
-				test: function(){}
-			});
-		}).not.toThrow();
-		expect(model).toBeDefined();
-		expect(typeof model.test).toBe('function');
-	});
-
+  it('should have a extend function', () => {
+    let model;
+    expect(() => {
+      model = Model.extend.call(ceres, {
+        test() {},
+      });
+    }).not.toThrow();
+    expect(model).toBeDefined();
+    expect(typeof model.test).toBe('function');
+  });
 });
