@@ -1,36 +1,35 @@
-var setupDirectory = require('../../src/setup/directory');
-var fs = require('fs');
+const fs = require('fs');
+const setupDirectory = require('../../src/setup/directory');
 
-var helpersDirectory = './spec/helpers/';
-var jsTest = /\.jsx?$/i;
+const helpersDirectory = './spec/helpers/';
+const jsTest = /\.jsx?$/i;
 
-describe('routes', function(){
-	var jsFiles = [];
-	beforeEach(function(){
-		jsFiles = fs.readdirSync(helpersDirectory).filter(function(file){
-			return jsTest.test(file);
-		});
-	});
+describe('routes', () => {
+  let jsFiles = [];
+  beforeEach(() => {
+    jsFiles = fs.readdirSync(helpersDirectory).filter(function(file) {
+      return jsTest.test(file);
+    });
+  });
 
-	it('should load any .js files in a directory and return an object of results', function(){
-		var result = setupDirectory(helpersDirectory);
-		expect(typeof result).toBe('object');
+  it('should load any .js files in a directory and return an object of results', () => {
+    const result = setupDirectory(helpersDirectory);
+    expect(typeof result).toBe('object');
 
-		jsFiles.forEach(function(filename){
-			// get the filename with the extension
-			var name = filename.replace(jsTest, '');
-			expect(result[name]).toBeDefined();
-		});
-	});
+    jsFiles.forEach(function(filename) {
+      // get the filename with the extension
+      const name = filename.replace(jsTest, '');
+      expect(result[name]).toBeDefined();
+    });
+  });
 
-
-	it('should run any .js files in a directory and call any functions', function(){
-		var result = setupDirectory(helpersDirectory, {
-			config: {
-				test: true
-			}
-		});
-		expect(result.TestController.test).toBe(true);
-		expect(typeof result.TestConfig).toBe('object');
-	});
+  it('should run any .js files in a directory and call any functions', () => {
+    const result = setupDirectory(helpersDirectory, {
+      config: {
+        test: true,
+      },
+    });
+    expect(result.TestController.test).toBe(true);
+    expect(typeof result.TestConfig).toBe('object');
+  });
 });
