@@ -1,22 +1,22 @@
-/*******************************************************************************
+/** *****************************************************************************
  * Ceres
  *
  * @author       Isaac Suttell <isaac_suttell@playstation.sony.com>
  * @file         Framework constructor
- ******************************************************************************/
+ ***************************************************************************** */
 
-var path = require('path');
-var Promise = require('bluebird');
-var EventEmitter = require('events');
-var cluster = require('cluster');
+const path = require('path');
+const Promise = require('bluebird');
+const EventEmitter = require('events');
+const cluster = require('cluster');
 
-var Config = require('./setup/Config');
-var setupCache = require('./setup/cache');
-var setupLogs = require('./setup/logs');
-var runStickyCluster = require('./setup/run-sticky-cluster');
-var runCluster = require('./setup/run-cluster');
-var runFork = require('./setup/run-fork');
-var Pid = require('./lib/Pid');
+const Config = require('./setup/Config');
+const setupCache = require('./setup/cache');
+const setupLogs = require('./setup/logs');
+const runStickyCluster = require('./setup/run-sticky-cluster');
+const runCluster = require('./setup/run-cluster');
+const runFork = require('./setup/run-fork');
+const Pid = require('./lib/Pid');
 
 function Ceres() {
   this.startTime = process.hrtime();
@@ -28,23 +28,24 @@ function Ceres() {
   this.Pipeline.create = this.Pipeline.create.bind(this);
 
   // Bind everything else just in case
-  for (var key in this) {
+  // eslint-disable-next-line no-restricted-syntax
+  for (const key in this) {
     if (typeof this[key] === 'function') {
       this[key] = this[key].bind(this);
     }
   }
 
   // Setup event emitter
-  this._events = new EventEmitter();
-  this.on = this._events.on.bind(this._events);
-  this.removeListener = this._events.removeListener.bind(this._events);
-  this.emit = this._events.emit.bind(this._events);
+  this.events = new EventEmitter();
+  this.on = this.events.on.bind(this.events);
+  this.removeListener = this.events.removeListener.bind(this.events);
+  this.emit = this.events.emit.bind(this.events);
 
-  this.on('configured', function(){
+  this.on('configured', () => {
     if (typeof this.config.hashids === 'object' && this.config.HashIds !== null) {
       this.HashIds = this.HashIds.call(this, this);
     }
-  }.bind(this));
+  });
 }
 
 /**
@@ -56,23 +57,23 @@ Ceres.prototype.Database = {};
 /**
  * Make contoller available at base level
  */
-Ceres.prototype.Controller = require(path.resolve(__dirname + '/controllers/Controller'));
+Ceres.prototype.Controller = require(path.resolve(`${__dirname}/controllers/Controller`));
 
 /**
  * Make the model available at the base level
  */
-Ceres.prototype.Model = require(path.resolve(__dirname + '/models/Model'));
+Ceres.prototype.Model = require(path.resolve(`${__dirname}/models/Model`));
 
 /**
  * Link to hashIds
  * @type    {Object}
  */
-Ceres.prototype.HashIds = require(path.resolve(__dirname + '/lib/hashIds'));
+Ceres.prototype.HashIds = require(path.resolve(`${__dirname}/lib/hashIds`));
 
 /**
  * Alias cut to Pipeline
  */
-Ceres.prototype.Pipeline = require(path.resolve(__dirname + '/render/Pipeline'));
+Ceres.prototype.Pipeline = require(path.resolve(`${__dirname}/render/Pipeline`));
 
 /**
  * Alias to run
@@ -82,46 +83,45 @@ Ceres.prototype.run = function run() {
   if (!this.config.secret) {
     throw new Error('Unable to find secret.');
   }
-  this.log._ceres.silly('Running application in %s mode', this.config.processManagement);
+  this.log.internal.silly('Running application in %s mode', this.config.processManagement);
   if (this.config.processManagement === 'fork') {
     return runFork.call(this, this);
-  } else if (this.config.processManagement === 'sticky-cluster') {
-    return runStickyCluster.call(this, this);
-  } else {
-    return runCluster.call(this, this);
   }
+  if (this.config.processManagement === 'sticky-cluster') {
+    return runStickyCluster.call(this, this);
+  }
+  return runCluster.call(this, this);
 };
 
 /**
  * Connect to database and cache
  * @return {Promise}
  */
-Ceres.prototype.connect = function() {
+Ceres.prototype.connect = function connect() {
   if (typeof this.config !== 'object') {
     Promise.reject(new Error('Ceres has not been configured yet'));
-    return;
-  } else if (this._databaseFactory) {
-    this.log._ceres.info('Using external database factory');
-    return new this._databaseFactory(this.config)
-      .then((results) => {
-        Object.assign(this, results);
-        return this;
-      });
+    return {};
+  }
+  if (this.DatabaseFactory) {
+    this.log.internal.info('Using external database factory');
+    return new this.DatabaseFactory(this.config).then(results => {
+      Object.assign(this, results);
+      return this;
+    });
   }
 
-  var type = this.config.db.type;
+  const { type } = this.config.db;
   if (['bookshelf', 'rethinkdb', 'mongodb'].indexOf(type) === -1) {
-    this.log._ceres.debug('Skipping database setup');
-    return setupCache(this)
-      .then(function(cache){
-        this.Cache = cache;
-        return this;
-      }.bind(this));
+    this.log.internal.debug('Skipping database setup');
+    return setupCache(this).then(cache => {
+      this.Cache = cache;
+      return this;
+    });
   }
 
-  this.log._ceres.silly('Connecting to %s...', type);
+  this.log.internal.silly('Connecting to %s...', type);
 
-  var databaseStartTime = Date.now();
+  const databaseStartTime = Date.now();
 
   // Expose these for any help function
   if (type === 'bookshelf') {
@@ -132,20 +132,26 @@ Ceres.prototype.connect = function() {
     this.Model.Mongodb = require('./models/types/MongodbModel');
   }
 
-  var connect = require(__dirname + '/db')(this.config, this);
+  const connection = require(`${__dirname}/db`)(this.config, this);
 
-  return connect.then(function(db){
-    var databaseStartupTime = Date.now() - databaseStartTime;
-    this.log._ceres.info('Connected to %s - %ss', type, (databaseStartupTime / 1000).toLocaleString(), {
-      duration: databaseStartupTime
-    });
-    this.Database = db;
-    return setupCache(this);
-  }.bind(this))
-    .then(function(cache){
+  return connection
+    .then(db => {
+      const databaseStartupTime = Date.now() - databaseStartTime;
+      this.log.internal.info(
+        'Connected to %s - %ss',
+        type,
+        (databaseStartupTime / 1000).toLocaleString(),
+        {
+          duration: databaseStartupTime,
+        }
+      );
+      this.Database = db;
+      return setupCache(this);
+    })
+    .then(cache => {
       this.Cache = cache;
       return this;
-    }.bind(this));
+    });
 };
 
 /**
@@ -153,8 +159,8 @@ Ceres.prototype.connect = function() {
  * @param  {Object} options External options
  * @return {Promise}
  */
-Ceres.prototype.configure = function(options) {
-  return new Promise(function(resolve, reject){
+Ceres.prototype.configure = function configure(options) {
+  return new Promise((resolve, reject) => {
     const startTime = Date.now();
     if (options instanceof Config) {
       this.config = options;
@@ -175,56 +181,73 @@ Ceres.prototype.configure = function(options) {
     this.log = this.logger();
 
     // Setup internal logger
-    this.log._ceres = setupLogs.init(this);
+    this.log.internal = setupLogs.init(this);
 
     // Check to see if this process is a child. Children do not need pid files as the parent handles that
-    this.isMaster = this.config.processManagement === 'fork' ? typeof process.env.CERES_UNIQUE_ID !== 'string' : cluster.isMaster;
+    this.isMaster =
+      this.config.processManagement === 'fork'
+        ? typeof process.env.CERES_UNIQUE_ID !== 'string'
+        : cluster.isMaster;
 
-    this.log._ceres.info('Starting %s...', this.config.name || 'ceres', {
+    this.log.internal.info('Starting %s...', this.config.name || 'ceres', {
       isMaster: this.isMaster,
-      pid: process.pid
+      pid: process.pid,
     });
-    this.log._ceres.info('Writing logs to %s', this.config.folders.logs);
+    this.log.internal.info('Writing logs to %s', this.config.folders.logs);
 
     // Fork mode handles this separatly
     if (this.config.processManagement !== 'fork') {
       // Log SIGTERM exit events
-      process.on('SIGTERM', (code) => {
-        this.log._ceres.info('%s process %s received SIGTERM; exiting...', this.isMaster ? 'Master' : 'Child', process.pid, {
-          isMaster: this.isMaster,
-          pid: process.pid
-        });
+      process.on('SIGTERM', code => {
+        this.log.internal.info(
+          '%s process %s received SIGTERM; exiting...',
+          this.isMaster ? 'Master' : 'Child',
+          process.pid,
+          {
+            isMaster: this.isMaster,
+            pid: process.pid,
+          }
+        );
         process.exit(code);
       });
     }
 
     this.emit('configured');
-    var duration = (Date.now() - startTime);
-    this.log._ceres.info('"%s" configuration loaded - %ss', this.config.env, (duration / 1000).toLocaleString(), { duration });
+    const duration = Date.now() - startTime;
+    this.log.internal.info(
+      '"%s" configuration loaded - %ss',
+      this.config.env,
+      (duration / 1000).toLocaleString(),
+      { duration }
+    );
 
     if (this.isMaster && this.config.pid && !options.disablePid) {
       // Setup Pid if we're configure
       this.pid = new Pid(this.config.pid);
-      this.pid.on('created', (pid) => {
-        this.log._ceres.info('Wrote pid to %s - %s', pid.options.path, pid.id);
+      this.pid.on('created', pid => {
+        this.log.internal.info('Wrote pid to %s - %s', pid.options.path, pid.id);
         resolve(this);
       });
       this.pid.on('existing', (existingPid, currentPid) => {
-        this.log._ceres.warn('Found an another active process at %s. Attempting to shut it down...', existingPid, {
+        this.log.internal.warn(
+          'Found an another active process at %s. Attempting to shut it down...',
           existingPid,
-          pid: currentPid
-        });
+          {
+            existingPid,
+            pid: currentPid,
+          }
+        );
       });
-      this.pid.on('removed', (pid) => {
-        this.log._ceres.debug('Process exiting. Removed pid file %s', pid.options.path, {
-          pid: pid.id
+      this.pid.on('removed', pid => {
+        this.log.internal.debug('Process exiting. Removed pid file %s', pid.options.path, {
+          pid: pid.id,
         });
       });
       this.pid.on('error', reject);
     } else {
       resolve(this);
     }
-  }.bind(this));
+  });
 };
 
 /**
@@ -233,8 +256,8 @@ Ceres.prototype.configure = function(options) {
  */
 function handleError(err) {
   // The logger may or may not be setup by the time this is called
-  if (this.log && typeof this.log._ceres === 'object') {
-    this.log._ceres.error(err);
+  if (this.log && typeof this.log.internal === 'object') {
+    this.log.internal.error(err);
   }
 
   // Always log to the stderr
@@ -247,8 +270,8 @@ function handleError(err) {
 /**
  * Store the database factory for later
  */
-Ceres.prototype.database = function(factory) {
-  this._databaseFactory = factory;
+Ceres.prototype.database = function database(factory) {
+  this.DatabaseFactory = factory;
   return this;
 };
 
@@ -257,21 +280,20 @@ Ceres.prototype.database = function(factory) {
  * @deprecated
  * @param  {Object} options
  */
-Ceres.prototype.load = function(options) {
-
+Ceres.prototype.load = function load(options) {
   // Load is typically used for scripts which do not need a pid by default
   if (typeof options.disablePid !== 'boolean') {
     options.disablePid = true;
   }
 
-  var instance = this;
+  const instance = this;
   return instance
     .configure(options)
     .then(instance.connect)
-    .then(function(ceres){
+    .then(ceres => {
       this.emit('before:run');
       return ceres;
-    }.bind(this))
+    })
     .catch(handleError.bind(this));
 };
 
@@ -279,14 +301,14 @@ Ceres.prototype.load = function(options) {
  * Load the app
  * @param  {Object} options
  */
-Ceres.prototype.exec = function(command, options) {
-  var instance = this;
+Ceres.prototype.exec = function exec(command, options) {
+  const instance = this;
   return instance
     .configure(options)
-    .then(function(ceres){
+    .then(ceres => {
       this.emit('before:run');
       return ceres;
-    }.bind(this))
+    })
     .then(command.bind(this, this))
     .catch(handleError.bind(this));
 };
