@@ -1,19 +1,22 @@
-var redis = require('redis');
-var Promise = require('bluebird');
+const redis = require('redis');
+const Promise = require('bluebird');
 
 /**
  * Setup cache instance
  * @param    {Objet}    options
  */
 function RedisCache(options, logger) {
-  this.options = Object.assign({
-    host: '127.0.0.1',
-    port: 6379,
-    ttl: 3600,
-    pass: '',
-    db: 0,
-    prefix: 'ceres'
-  }, options);
+  this.options = Object.assign(
+    {
+      host: '127.0.0.1',
+      port: 6379,
+      ttl: 3600,
+      pass: '',
+      db: 0,
+      prefix: 'ceres',
+    },
+    options
+  );
 
   /**
    * Save logger
@@ -29,7 +32,7 @@ function RedisCache(options, logger) {
   /**
    * Log any errors
    */
-  this.client.on('error', function(err){
+  this.client.on('error', err => {
     if (this.logger) {
       this.logger.error(err);
     } else {
@@ -59,17 +62,17 @@ RedisCache.prototype.get = function get(key) {
     this.logger.silly('Getting %s', key);
   }
 
-  return new Promise(function(resolve, reject){
-    this.client.get(key, function(err, reply){
-      if(err) {
+  return new Promise((resolve, reject) => {
+    this.client.get(key, (err, reply) => {
+      if (err) {
         reject(err);
-      } else if(typeof reply !== 'string') {
+      } else if (typeof reply !== 'string') {
         resolve();
       } else {
         resolve(JSON.parse(reply.toString()));
       }
     });
-  }.bind(this));
+  });
 };
 
 /**
@@ -84,17 +87,19 @@ RedisCache.prototype.keys = function keys(search) {
     this.logger.silly('Gettings keysL %s', search);
   }
 
-  return new Promise(function(resolve, reject){
-    this.client.keys(search, function(err, reply){
-      if(err) {
+  return new Promise((resolve, reject) => {
+    this.client.keys(search, (err, reply) => {
+      if (err) {
         reject(err);
       } else {
-        resolve(reply.map(function(item){
-          return JSON.parse(item.toString());
-        }));
+        resolve(
+          reply.map(item => {
+            return JSON.parse(item.toString());
+          })
+        );
       }
     });
-  }.bind(this));
+  });
 };
 
 /**
@@ -105,16 +110,19 @@ RedisCache.prototype.set = function set(key, body, options) {
     return Promise.reject(new TypeError('key is not a string'));
   }
 
-  options = Object.assign({
-    expires: 60 * 60 * 12
-  }, options);
+  options = Object.assign(
+    {
+      expires: 60 * 60 * 12,
+    },
+    options
+  );
 
   if (this.logger) {
     this.logger.silly('Setting %s to', key, body);
   }
 
-  return new Promise(function(resolve, reject){
-    this.client.set(key, JSON.stringify(body), function(err){
+  return new Promise((resolve, reject) => {
+    this.client.set(key, JSON.stringify(body), err => {
       if (err) {
         reject(err);
         return;
@@ -130,15 +138,15 @@ RedisCache.prototype.set = function set(key, body, options) {
         this.logger.silly('Expiring %s after %ss', key, options.expires);
       }
 
-      this.client.expire(key, options.expires, function(e){
+      this.client.expire(key, options.expires, e => {
         if (e) {
           reject(e);
         } else {
           resolve(body);
         }
       });
-    }.bind(this));
-  }.bind(this));
+    });
+  });
 };
 
 /**
@@ -153,15 +161,15 @@ RedisCache.prototype.del = function del(key) {
     this.logger.silly('Deleting %s to', key);
   }
 
-  return new Promise(function(resolve, reject){
-    this.client.del(key, function(err){
+  return new Promise((resolve, reject) => {
+    this.client.del(key, err => {
       if (err) {
         reject(err);
         return;
       }
       resolve();
     });
-  }.bind(this));
+  });
 };
 
 module.exports = RedisCache;
