@@ -1,42 +1,50 @@
-/*******************************************************************************
+/** *****************************************************************************
  * Responses
  *
  * @author       Isaac Suttell <isaac_suttell@playstation.sony.com>
  * @file         Default rest responses
- ******************************************************************************/
+ ***************************************************************************** */
 
 /**
  * HTTP Status Codes
  *
  * @type    {Object}
  */
-var STATUS = {
+const STATUS = {
   OK: 200,
   NO_CONTENT: 204,
   BAD_REQUEST: 400,
   FORBIDDEN: 401,
   PERMISSION_DENIED: 403,
   NOT_FOUND: 404,
-  ERROR: 500
+  ERROR: 500,
 };
 
-var Responses = {
+const Responses = {
   /**
    * OK Response
    * @param     {Mixed}    data
    */
-  send: function(data) {
+  send(data) {
     if (this.res.headersSent || this.res.finished) {
-      this.log.warn('%s %s', this.req.method, this.req.originalUrl, new Error('Headers already sent'));
+      this.log.warn(
+        '%s %s',
+        this.req.method,
+        this.req.originalUrl,
+        new Error('Headers already sent')
+      );
       return;
     }
-    this.res.status(STATUS.OK).json(data).end();
+    this.res
+      .status(STATUS.OK)
+      .json(data)
+      .end();
   },
 
   /**
    * Nothing to send pack but we succeed
    */
-  noContent: function() {
+  noContent() {
     this.res.status(STATUS.NO_CONTENT).end();
   },
 
@@ -44,8 +52,8 @@ var Responses = {
    * Can't find it
    * @param     {String}    context
    */
-  notFound: function(context) {
-    var err = new Error('Not Found' + (typeof context === 'string' ? ': ' + context: ''));
+  notFound(context) {
+    const err = new Error(`Not Found${typeof context === 'string' ? `: ${context}` : ''}`);
     err.status = STATUS.NOT_FOUND;
     this.fail(err);
   },
@@ -54,8 +62,8 @@ var Responses = {
    * User doesn't have access
    * @param     {String}    context
    */
-  forbidden: function(context) {
-    var err = new Error('Forbidden' + (typeof context === 'string' ? ': ' + context: ''));
+  forbidden(context) {
+    const err = new Error(`Forbidden${typeof context === 'string' ? `: ${context}` : ''}`);
     err.status = STATUS.FORBIDDEN;
     this.fail(err);
   },
@@ -64,8 +72,8 @@ var Responses = {
    * User doesn't have access
    * @param     {String}    context
    */
-  permissionDenied: function(context) {
-    var err = new Error('Forbidden' + (typeof context === 'string' ? ': ' + context: ''));
+  permissionDenied(context) {
+    const err = new Error(`Forbidden${typeof context === 'string' ? `: ${context}` : ''}`);
     err.status = STATUS.PERMISSION_DENIED;
     this.fail(err);
   },
@@ -74,8 +82,8 @@ var Responses = {
    * Client sent a request that we can't process for some reason
    * @param     {String}    context
    */
-  badRequest: function(context) {
-    var err = new Error('Bad Request' + (typeof context === 'string' ? ': ' + context: ''));
+  badRequest(context) {
+    const err = new Error(`Bad Request${typeof context === 'string' ? `: ${context}` : ''}`);
     err.status = STATUS.BAD_REQUEST;
     this.fail(err);
   },
@@ -84,11 +92,10 @@ var Responses = {
    * There was an error!!!
    * @param     {Mixed}    err
    */
-  fail: function(err) {
+  fail(err) {
     // Throw and exit the call stack or promise chain
     this.next(err);
-  }
-
+  },
 };
 
 module.exports = Responses;
