@@ -1,5 +1,5 @@
-var express = require('express');
-var wrapRoute = require('./wrapRoute');
+const express = require('express');
+const wrapRoute = require('./wrapRoute');
 
 /**
  * Get the full path of a controller endpoint
@@ -8,7 +8,7 @@ var wrapRoute = require('./wrapRoute');
  * @return {String}
  */
 function getFullPath(controller, path) {
-  return (controller.endpoint + path).replace('//','/');
+  return (controller.endpoint + path).replace('//', '/');
 }
 
 /**
@@ -20,10 +20,11 @@ function getFunctionName(val) {
   if (typeof val === 'string') {
     // Default
     return val;
-  } else if (typeof val === 'function') {
+  }
+  if (typeof val === 'function') {
     return val.constructor.name;
   }
-  return;
+  return undefined;
 }
 
 /**
@@ -35,11 +36,12 @@ function getFunctionName(val) {
 function getHandler(controller, lastArg) {
   if (typeof lastArg === 'function') {
     return lastArg;
-  } else if (typeof lastArg === 'string') {
+  }
+  if (typeof lastArg === 'string') {
     // By default look for a method with the fnName on `this`
     return controller[lastArg];
   }
-  return;
+  return undefined;
 }
 
 /**
@@ -50,15 +52,21 @@ function getHandler(controller, lastArg) {
  * @return   {Array<Function>}
  */
 function getMiddleware(controller, ceres, route) {
-  var args = [];
+  let args = [];
   // Controller Middleware
   if (controller.middleware instanceof Array) {
     args = args.concat(controller.middleware);
   } else if (typeof controller.middleware === 'function') {
     // Run it once so we can configure it. It should return a function
-    var userMiddleware = controller.middleware.call(controller, ceres.config.middleware, ceres.config);
+    const userMiddleware = controller.middleware.call(
+      controller,
+      ceres.config.middleware,
+      ceres.config
+    );
 
-    args = args.concat(userMiddleware instanceof Array !== true ? [userMiddleware] : userMiddleware);
+    args = args.concat(
+      userMiddleware instanceof Array !== true ? [userMiddleware] : userMiddleware
+    );
   }
 
   // Attach another other user supplied middleware
@@ -79,9 +87,8 @@ function getLastArg(controller, route) {
   if (controller.routes[route] instanceof Array) {
     // Array option so users can inject middleware
     return controller.routes[route].pop();
-  } else {
-    return controller.routes[route];
   }
+  return controller.routes[route];
 }
 
 /**
@@ -91,27 +98,27 @@ function getLastArg(controller, route) {
  */
 function parseRouteString(route) {
   // Parse
-  var parts = route.split(' ');
+  const parts = route.split(' ');
 
-  if(parts.length !== 2) {
+  if (parts.length !== 2) {
     throw new Error('Invalid path string');
   }
 
   /**
-	 * HTTP method
-	 * @type {String}
-	 */
-  var method = parts[0].toLowerCase();
+   * HTTP method
+   * @type {String}
+   */
+  const method = parts[0].toLowerCase();
 
   /**
-	 * The sub path or tail end of a path
-	 * @type {String}
-	 */
-  var path = parts[1];
+   * The sub path or tail end of a path
+   * @type {String}
+   */
+  const path = parts[1];
 
   return {
-    method: method,
-    path: path
+    method,
+    path,
   };
 }
 
@@ -121,73 +128,69 @@ function parseRouteString(route) {
  * @param    {Ceres}         ceres
  */
 function controllerRoutes(controller, ceres) {
-  if (typeof controller.routes !== 'object'){
+  if (typeof controller.routes !== 'object') {
     throw Error('Missing controller.routes');
   }
 
   /**
-	 * Final results
-	 * @type    {Array<Object>}
-	 */
-  var routes = [];
+   * Final results
+   * @type    {Array<Object>}
+   */
+  const routes = [];
 
   // Loop through all of them
-  for (var route in controller.routes) {
-    if (!controller.routes.hasOwnProperty(route)) {
-      continue;
-    }
-
+  Object.keys(controller.routes).forEach(route => {
     // Parse
-    var parts = parseRouteString(route);
+    const parts = parseRouteString(route);
 
     /**
-		 * HTTP method
-		 * @type {String}
-		 */
-    var method = parts.method;
+     * HTTP method
+     * @type {String}
+     */
+    const { method } = parts;
 
     /**
-		 * The sub path or tail end of a path
-		 * @type {String}
-		 */
-    var path = parts.path;
+     * The sub path or tail end of a path
+     * @type {String}
+     */
+    const { path } = parts;
 
     /**
-		 * The complete path from the root url
-		 * @type {String}
-		 */
-    var fullPath = getFullPath(controller, path);
+     * The complete path from the root url
+     * @type {String}
+     */
+    const fullPath = getFullPath(controller, path);
 
     /**
-		 * Get the last argument which is the main method
-		 * @type    {Mixed}
-		 */
-    var lastArg = getLastArg(controller, route);
+     * Get the last argument which is the main method
+     * @type    {Mixed}
+     */
+    const lastArg = getLastArg(controller, route);
 
     /**
-		 * Get fn name aka value
-		 * @type    {String}
-		 */
-    var fnName = getFunctionName(lastArg);
+     * Get fn name aka value
+     * @type    {String}
+     */
+    const fnName = getFunctionName(lastArg);
 
     /**
-	 * Fet actual fn
-	 * @type    {Function}
-	 */
-    var handler = getHandler(controller, lastArg);
+     * Fet actual fn
+     * @type    {Function}
+     */
+    const handler = getHandler(controller, lastArg);
 
     if (typeof handler !== 'function') {
-      throw new Error(fullPath + ' - ' + controller.name + '.' + fnName + ' is not a function');
+      throw new Error(`${fullPath} - ${controller.name}.${fnName} is not a function`);
     }
 
     /**
-		 * Setup middleware for controller and route
-		 * @type    {Array<Functions>}
-		 */
-    var args = getMiddleware(controller, ceres, route);
+     * Setup middleware for controller and route
+     * @type    {Array<Functions>}
+     */
+    const args = getMiddleware(controller, ceres, route);
 
     // Ensure middleware functons are valid
-    args.forEach(function(arg){
+    args.forEach(arg => {
       if (typeof arg !== 'function') {
         throw new Error('middleware is not a function');
       }
@@ -197,28 +200,38 @@ function controllerRoutes(controller, ceres) {
     args.unshift(path);
 
     /**
-		 * Bind custom this context to route
-		 * @type    {Function}
-		 */
-    var fn = wrapRoute(handler, controller, ceres);
+     * Bind custom this context to route
+     * @type    {Function}
+     */
+    const fn = wrapRoute(handler, controller, ceres);
 
     // Add to args at the end
     args.push(fn);
 
     // Ensure the method exists in Express
     if (typeof express.Router[method] !== 'function') {
-      throw new Error(controller.name + ' - Ignoring ' + method.toUpperCase() + ' ' + fullPath + ': router.' + method + ' is not a function');
+      throw new Error(
+        `${
+          controller.name
+        } - Ignoring ${method.toUpperCase()} ${fullPath}: router.${method} is not a function`
+      );
     }
 
     // Save
     routes.push({
-      method: method,
-      args: args
+      method,
+      args,
     });
 
     // Log
-    ceres.log._ceres.silly('%s - Setting up %s %s - %s', controller.name, method.toUpperCase(), fullPath, fnName);
-  }
+    ceres.log.internal.silly(
+      '%s - Setting up %s %s - %s',
+      controller.name,
+      method.toUpperCase(),
+      fullPath,
+      fnName
+    );
+  });
 
   return routes;
 }
@@ -229,12 +242,12 @@ function controllerRoutes(controller, ceres) {
  * @return   {Express.Router}
  */
 module.exports = function controllerRouter(ceres) {
-  var router = new express.Router({
-    mergeParams: true
+  const router = new express.Router({
+    mergeParams: true,
   });
-  var routes = controllerRoutes(this, ceres);
-  routes.forEach(function(route){
-    router[route.method].apply(router, route.args);
+  const routes = controllerRoutes(this, ceres);
+  routes.forEach(route => {
+    router[route.method](...route.args);
   });
   return router;
 };
