@@ -1,18 +1,20 @@
-var Hashids = require('hashids');
+const Hashids = require('hashids');
 
-module.exports = function(Ceres) {
+module.exports = function ids(Ceres) {
   if (typeof Ceres.config.hashIds !== 'object') {
-    var setupError = function() {
+    const setupError = () => {
       throw new Error('hashIds not configured');
     };
 
     return {
       decode: setupError,
-      encode: setupError
+      encode: setupError,
     };
-  } else {
-    Ceres.log._ceres.silly('HashIds enabled: Using secret "%s" for ids', Ceres.config.hashIds.secret);
-
-    return new Hashids(Ceres.config.hashIds.secret, Ceres.config.hashIds.minLength);
   }
+  Ceres.log.internal.silly(
+    'HashIds enabled: Using secret "%s" for ids',
+    Ceres.config.hashIds.secret
+  );
+
+  return new Hashids(Ceres.config.hashIds.secret, Ceres.config.hashIds.minLength);
 };
