@@ -1,6 +1,6 @@
-module.exports = function(ceres) {
+module.exports = ceres => {
   if (typeof ceres.config.hashIds !== 'object') {
-    return function(req, res, next) {
+    return (req, res, next) => {
       next(new Error('hashIds no enabled'));
     };
   }
@@ -12,10 +12,10 @@ module.exports = function(ceres) {
    * @param  {Express.Response}   res
    * @param  {Function} next
    */
-  return function(req, res, next){
+  return (req, res, next) => {
     if (req.params.key) {
-      var id = ceres.HashIds.decode(req.params.key)[0];
-      if (isNaN(id)) {
+      const id = ceres.HashIds.decode(req.params.key)[0];
+      if (Number.isNaN(id)) {
         ceres.log.warn('Error decoding key', req.params.key);
         next(new Error('InvalidKey'));
         return;
