@@ -1,5 +1,6 @@
-module.exports = function() {
-  return function(req, res, next){ // eslint-disable-line no-unused-vars
+module.exports = () => {
+  return (req, res, next) => {
+    // eslint-disable-line no-unused-vars
     next(new Error('Not Found: Unable to find resource'));
   };
 };
