@@ -1,14 +1,16 @@
-var webpackDevMiddleware = require('webpack-dev-middleware');
-var webpack = require('webpack');
+// eslint-disable-next-line import/no-extraneous-dependencies
+const webpackDevMiddleware = require('webpack-dev-middleware');
+// eslint-disable-next-line import/no-extraneous-dependencies
+const webpack = require('webpack');
 
-module.exports = function(ceres) {
-  var config = ceres.config.webpackConfig;
+module.exports = ceres => {
+  const { webpackConfig: config } = ceres.config;
   config.output.path = '/';
-  var compiler = webpack(config);
+  const compiler = webpack(config);
   return webpackDevMiddleware(compiler, {
     noInfo: false,
     quiet: false,
     publicPath: '/assets',
-    stats: config.stats
+    stats: config.stats,
   });
 };
