@@ -1,9 +1,9 @@
-/*******************************************************************************
+/** *****************************************************************************
  * Base Model
- ******************************************************************************/
+ ***************************************************************************** */
 
-var path = require('path');
-var pkg = require('../../package.json');
+const path = require('path');
+const pkg = require('../../package.json');
 
 /**
  * Setup a consistent model system that interacts with an database
@@ -13,16 +13,16 @@ function Model(Ceres, props) {
   this.version = pkg.version;
 
   // Import ORM
-  var model;
+  let model;
 
-  var type = props.type || Ceres.config.db.type;
+  const type = props.type || Ceres.config.db.type;
 
   if (['bookshelf', 'rethinkdb', 'mongodb'].indexOf(type.toLowerCase()) > -1) {
-    var filename = type.slice(0, 1).toUpperCase() + type.slice(1) + 'Model.js';
-    var orm = require(path.resolve(__dirname, './types', filename));
+    const filename = `${type.slice(0, 1).toUpperCase() + type.slice(1)}Model.js`;
+    const orm = require(path.resolve(__dirname, './types', filename));
     model = orm.extend.call(Ceres, props);
   } else {
-    throw new Error('Unknown model type: ' + type);
+    throw new Error(`Unknown model type: ${type}`);
   }
 
   // Pull everything off and into this instance
