@@ -1,12 +1,12 @@
-/*******************************************************************************
+/** *****************************************************************************
  * Base Model
- ******************************************************************************/
+ ***************************************************************************** */
 
-var _ = require('lodash');
-var moment = require('moment');
-var Promise = require('bluebird');
-var assertNotNull = require('../../lib/assert').assertNotNull;
-var assertDefined = require('../../lib/assert').assertDefined;
+const _ = require('lodash');
+const moment = require('moment');
+const Promise = require('bluebird');
+const { assertNotNull } = require('../../lib/assert');
+const { assertDefined } = require('../../lib/assert');
 
 /**
  * Setup
@@ -24,15 +24,17 @@ function BookshelfModel(props) {
 
   // Make raw function from knex available directly on the model
   this.raw = this.database.knex.raw;
-
   // Make bookself methods available directly on our model so we don't have to
   // do `this.model.model[fn]` to access them.
-  for (var fnName in this.model) {
+  // eslint-disable-next-line no-restricted-syntax
+  for (const fnName in this.model) {
+    // eslint-disable-next-line no-prototype-builtins
     if (!this.model.hasOwnProperty(fnName)) {
+      // eslint-disable-next-line no-continue
       continue;
     } else if (typeof this.model[fnName] === 'function' && typeof this[fnName] !== 'undefined') {
       // Ensure we don't override anything
-      throw new Error('bookshelf.' + fnName + ' conflicts with local model');
+      throw new Error(`bookshelf.${fnName} conflicts with local model`);
     } else if (typeof this.model[fnName] === 'function') {
       // Gotta bind to ensure the right this
       this[fnName] = this.model[fnName].bind(this.model);
@@ -48,13 +50,15 @@ function BookshelfModel(props) {
  */
 BookshelfModel.prototype.create = function create(body) {
   assertNotNull(this.model);
-  return new this.model(body).save(null, { // eslint-disable-line new-cap
-    method: 'insert'
-  })
-    .then(function(model){
-    // Look up any relations
+  // eslint-disable-next-line new-cap
+  return new this.model(body)
+    .save(null, {
+      method: 'insert',
+    })
+    .then(model => {
+      // Look up any relations
       return this.read(model.id);
-    }.bind(this));
+    });
 };
 
 /**
@@ -68,15 +72,17 @@ BookshelfModel.prototype.read = function read(id) {
   assertDefined(id, 'id');
   if (id instanceof Array) {
     return this.model.where('id', 'IN', id).fetchAll(this.fetch);
-  } else if (_.isObject(id)) {
-    return new this.model({ // eslint-disable-line new-cap
-      id: id.id
-    }).fetch(this.fetch);
-  } else {
-    return new this.model({ // eslint-disable-line new-cap
-      id: id
+  }
+  if (_.isObject(id)) {
+    // eslint-disable-next-line new-cap
+    return new this.model({
+      id: id.id,
     }).fetch(this.fetch);
   }
+  // eslint-disable-next-line new-cap
+  return new this.model({
+    id,
+  }).fetch(this.fetch);
 };
 
 /**
@@ -96,7 +102,8 @@ BookshelfModel.prototype.readAll = function readAll() {
  */
 BookshelfModel.prototype.find = function find(query) {
   assertNotNull(this.model);
-  return new this.model(query).fetch(this.fetch); // eslint-disable-line new-cap
+  // eslint-disable-next-line new-cap
+  return new this.model(query).fetch(this.fetch);
 };
 
 /**
@@ -107,7 +114,7 @@ BookshelfModel.prototype.find = function find(query) {
  */
 BookshelfModel.prototype.update = function update(body, id) {
   if (typeof id === 'undefined' && typeof body.id !== 'undefined') {
-    id = body.id;
+    ({ id } = body);
   }
   assertNotNull(this.model);
   assertDefined(id, 'id');
@@ -119,15 +126,18 @@ BookshelfModel.prototype.update = function update(body, id) {
   delete body.created_at; // You can only create it once
   delete body.updated_at; // Handled by DB
 
-  return new this.model({ // eslint-disable-line new-cap
-    id: id
-  }).save(body, {
-    patch: true,
-    method: 'update'
-  }).then(function() {
-    // Get relations
-    return this.read(id);
-  }.bind(this));
+  // eslint-disable-next-line new-cap
+  return new this.model({
+    id,
+  })
+    .save(body, {
+      patch: true,
+      method: 'update',
+    })
+    .then(() => {
+      // Get relations
+      return this.read(id);
+    });
 };
 
 /**
@@ -139,9 +149,11 @@ BookshelfModel.prototype.updateAll = function updateAll(body) {
   if (body instanceof Array !== true) {
     body = [body];
   }
-  return Promise.all(body.map(function(doc) {
-    return this.update(doc, doc.id);
-  }.bind(this)));
+  return Promise.all(
+    body.map(doc => {
+      return this.update(doc, doc.id);
+    })
+  );
 };
 
 /**
@@ -153,8 +165,9 @@ BookshelfModel.prototype.updateAll = function updateAll(body) {
 BookshelfModel.prototype.del = function del(id) {
   assertNotNull(this.model);
   assertDefined(id, 'id');
-  return new this.model({ // eslint-disable-line new-cap
-    id: id
+  // eslint-disable-next-line new-cap
+  return new this.model({
+    id,
   }).destroy();
 };
 
@@ -184,18 +197,22 @@ module.exports.convertTimestampsToUnix = function convertTimestampsToUnix(fields
    *
    * @param     {Array}    attrs
    */
-  return function(attrs) {
-    return _.reduce(attrs, function(memo, val, key) {
-      if (options.camelCase) {
-        key = _.camelCase(key);
-      }
-      if (val && fields.indexOf(key) > -1) {
-        memo[key] = moment(val).format('x');
-      } else {
-        memo[key] = val;
-      }
-      return memo;
-    }, {});
+  return attrs => {
+    return _.reduce(
+      attrs,
+      (memo, val, key) => {
+        if (options.camelCase) {
+          key = _.camelCase(key);
+        }
+        if (val && fields.indexOf(key) > -1) {
+          memo[key] = moment(val).format('x');
+        } else {
+          memo[key] = val;
+        }
+        return memo;
+      },
+      {}
+    );
   };
 };
 
@@ -213,17 +230,23 @@ module.exports.convertDatesToISO8601 = function convertDatesToISO8601(fields, op
    *
    * @param     {Array}    attrs
    */
-  return function(attrs) {
-    return _.reduce(attrs, function(memo, val, key) {
-      if (options.camelCase) {
-        key = _.camelCase(key);
-      }
-      if (val && fields.indexOf(key) > -1) {
-        memo[key] = moment(val).utc().format();
-      } else {
-        memo[key] = val;
-      }
-      return memo;
-    }, {});
+  return attrs => {
+    return _.reduce(
+      attrs,
+      (memo, val, key) => {
+        if (options.camelCase) {
+          key = _.camelCase(key);
+        }
+        if (val && fields.indexOf(key) > -1) {
+          memo[key] = moment(val)
+            .utc()
+            .format();
+        } else {
+          memo[key] = val;
+        }
+        return memo;
+      },
+      {}
+    );
   };
 };
