@@ -1,9 +1,9 @@
-/*******************************************************************************
+/** *****************************************************************************
  * Base Model
- ******************************************************************************/
+ ***************************************************************************** */
 
-var _ = require('lodash');
-var Promise = require('bluebird');
+const _ = require('lodash');
+const Promise = require('bluebird');
 
 /**
  * Initial model
@@ -27,16 +27,16 @@ MongodbModel.prototype = Object.assign(MongodbModel.prototype, {
    * @param     {Object}    body
    * @return    {promise}
    */
-  create: function(body) {
-    return new Promise(function(resolve, reject){
-      this.collection.insert(body, function(err, result){
-        if(err) {
+  create(body) {
+    return new Promise((resolve, reject) => {
+      this.collection.insert(body, (err, result) => {
+        if (err) {
           reject(err);
         } else {
           resolve(result.ops[0]);
         }
       });
-    }.bind(this));
+    });
   },
 
   /**
@@ -45,37 +45,40 @@ MongodbModel.prototype = Object.assign(MongodbModel.prototype, {
    * @param     {Mixed}    id
    * @return    {promise}
    */
-  read: function(id) {
+  read(id) {
     if (typeof id !== 'string') {
       return Promise.reject(new Error('id is not a string'));
     }
-    return new Promise(function(resolve, reject){
-      this.collection.findOne({
-        _id: id
-      }, function(err, result){
-        if(err) {
-          reject(err);
-        } else {
-          resolve(result);
+    return new Promise((resolve, reject) => {
+      this.collection.findOne(
+        {
+          _id: id,
+        },
+        (err, result) => {
+          if (err) {
+            reject(err);
+          } else {
+            resolve(result);
+          }
         }
-      });
-    }.bind(this));
+      );
+    });
   },
 
   /**
    * Return all records in collection
    * @return {Promise}
    */
-  readAll: function() {
-    return new Promise(function(resolve, reject){
-      this.collection.find({}, function(err, result){
+  readAll() {
+    return new Promise((resolve, reject) => {
+      this.collection.find({}, (err, result) => {
         if (err) {
           reject(err);
         } else {
           resolve(result.toArray());
         }
       });
-    }.bind(this));
+    });
   },
 
   /**
@@ -84,19 +87,19 @@ MongodbModel.prototype = Object.assign(MongodbModel.prototype, {
    * @param     {Mixed}    id
    * @return    {promise}
    */
-  find: function(query) {
+  find(query) {
     if (typeof query !== 'object') {
       return Promise.reject(new Error('query is not an object'));
     }
-    return new Promise(function(resolve, reject){
-      this.collection.find(query, function(err, result){
-        if(err) {
+    return new Promise((resolve, reject) => {
+      this.collection.find(query, (err, result) => {
+        if (err) {
           reject(err);
         } else {
           resolve(result.toArray());
         }
       });
-    }.bind(this));
+    });
   },
 
   /**
@@ -105,22 +108,28 @@ MongodbModel.prototype = Object.assign(MongodbModel.prototype, {
    * @param     {Object}    body
    * @return    {promise}
    */
-  update: function(body, id) {
+  update(body, id) {
     if (!id) {
+      // eslint-disable-next-line no-underscore-dangle
       id = body._id;
     }
+    // eslint-disable-next-line no-underscore-dangle
     delete body._id; // Can't update the ID
-    return new Promise(function(resolve, reject){
-      this.collection.updateOne({
-        _id: id
-      }, body, function(err, result){
-        if(err) {
-          reject(err);
-        } else {
-          resolve(result);
+    return new Promise((resolve, reject) => {
+      this.collection.updateOne(
+        {
+          _id: id,
+        },
+        body,
+        (err, result) => {
+          if (err) {
+            reject(err);
+          } else {
+            resolve(result);
+          }
         }
-      });
-    }.bind(this));
+      );
+    });
   },
 
   /**
@@ -129,19 +138,22 @@ MongodbModel.prototype = Object.assign(MongodbModel.prototype, {
    * @param     {Number}    id
    * @return    {promise}
    */
-  del: function(id) {
-    return new Promise(function(resolve, reject){
-      this.collection.remove({
-        _id: id
-      }, function(err, result){
-        if(err) {
-          reject(err);
-        } else {
-          resolve(result);
+  del(id) {
+    return new Promise((resolve, reject) => {
+      this.collection.remove(
+        {
+          _id: id,
+        },
+        (err, result) => {
+          if (err) {
+            reject(err);
+          } else {
+            resolve(result);
+          }
         }
-      });
-    }.bind(this));
-  }
+      );
+    });
+  },
 });
 
 /**
