@@ -1,22 +1,21 @@
-/*******************************************************************************
+/** *****************************************************************************
  * Pipeline
  *
  * @author       Isaac Suttell <isaac_suttell@playstation.sony.com>
  * @file         Render Workflow to hepl ensure we consistently produce the
  *               same results across each page and prints
  * @flow
- ******************************************************************************/
+ ***************************************************************************** */
 
-module.exports = function() {
-
+module.exports = () => {
   function AssetFactory(assets) {
     return {
-      forPayload: function() {
+      forPayload() {
         return {
           js: assets.js,
-          css: assets.css
+          css: assets.css,
         };
-      }
+      },
     };
   }
 
