@@ -1,27 +1,27 @@
-var Promise = require('bluebird');
-var fork = require('child_process').fork;
+const Promise = require('bluebird');
+const { fork } = require('child_process');
 
-var Server = require('./Server');
-var logStartTime = require('../lib/logStartTime');
+const Server = require('./Server');
+const logStartTime = require('../lib/logStartTime');
 
 /**
  * Unique id of each child process
  * @type    {Number}
  */
-var uniqueIds = 0;
+let uniqueIds = 0;
 
 /**
  * Active child processes
  * @type    {Array}
  */
-var workers = [];
+const workers = [];
 
 /**
  * Settings to control worker
  * @type    {Object}
  */
-var childSettings = {
-  restart: true
+const childSettings = {
+  restart: true,
 };
 
 /**
@@ -33,15 +33,15 @@ var childSettings = {
  */
 function spawn(ceres, port, workerIndex) {
   /**
-	 * Give each child a uniqueId
-	 * @type    {Number}
-	 */
-  var id = ++uniqueIds;
+   * Give each child a uniqueId
+   * @type    {Number}
+   */
+  const id = uniqueIds;
+  uniqueIds += 1;
 
   // Rerun with the same arguments
-  var worker = fork(process.argv[1], process.argv, {
+  const worker = fork(process.argv[1], process.argv, {
     env: {
-
       // Each process gets a unique port
       PORT: port,
 
@@ -49,38 +49,38 @@ function spawn(ceres, port, workerIndex) {
       CERES_UNIQUE_ID: id,
 
       // Let the children know their index
-      WORKER_INDEX: workerIndex
-    }
+      WORKER_INDEX: workerIndex,
+    },
   });
 
   // Loggin
-  ceres.log._ceres.debug('Spawned child #%s - %s', id, worker.pid);
+  ceres.log.internal.debug('Spawned child #%s - %s', id, worker.pid);
 
   // Keep track of the active workers
   workers.push(worker);
 
   // Log any errors
-  worker.on('error', function(err){
-    ceres.log._ceres.error(err);
+  worker.on('error', err => {
+    ceres.log.internal.error(err);
   });
 
   // Listen to messages from the parent
-  worker.on('message', function(obj){
-    ceres.log._ceres.error('%s recieved %s', worker.pid, obj);
+  worker.on('message', obj => {
+    ceres.log.internal.error('%s recieved %s', worker.pid, obj);
     if (typeof obj === 'object') {
       Object.assign(childSettings, obj);
     }
   });
 
   // Attempt to respawn unexcepted crashes
-  worker.on('exit', function(code, signal){
+  worker.on('exit', (code, signal) => {
     if (code && code > 0) {
-      ceres.log._ceres.error('%s exited with %s', worker.pid, code);
+      ceres.log.internal.error('%s exited with %s', worker.pid, code);
     } else if (signal) {
-      ceres.log._ceres.debug('%s received %s', worker.pid, signal);
+      ceres.log.internal.debug('%s received %s', worker.pid, signal);
     }
 
-    var index = workers.indexOf(worker);
+    const index = workers.indexOf(worker);
     if (index !== -1) {
       workers.splice(index, 1);
     }
@@ -88,7 +88,7 @@ function spawn(ceres, port, workerIndex) {
     // Respawn the child if we get an error code. Do not respawn if we received
     // a signal. Respawning on signals leads to endless recurision. It hurts.
     if (code && code > 0 && childSettings.restart === true) {
-      ceres.log._ceres.info('Repawning new process...', worker.pid);
+      ceres.log.internal.info('Repawning new process...', worker.pid);
       spawn(ceres, port, workerIndex);
     }
   });
@@ -100,14 +100,19 @@ function spawn(ceres, port, workerIndex) {
  * @return {Promise}
  */
 function listen(ceres) {
-  return new Promise(function(resolve, reject){
+  return new Promise((resolve, reject) => {
     try {
-      Server.call(ceres, ceres).listen(ceres.config.port, function(){
+      Server.call(ceres, ceres).listen(ceres.config.port, () => {
         logStartTime('Child took %ds to start listening', ceres);
-        ceres.log._ceres.info('Child #%s listening on %d (%s)', process.env.CERES_UNIQUE_ID, ceres.config.port, ceres.config.env);
+        ceres.log.internal.info(
+          'Child #%s listening on %d (%s)',
+          process.env.CERES_UNIQUE_ID,
+          ceres.config.port,
+          ceres.config.env
+        );
         resolve();
       });
-    } catch(err) {
+    } catch (err) {
       reject(err);
     }
   });
@@ -118,31 +123,36 @@ function listen(ceres) {
  * @param  {Ceres}    ceres
  * @return {Promise}
  */
-module.exports = function(ceres) {
-  return new Promise(function(resolve, reject){
+module.exports = ceres => {
+  return new Promise((resolve, reject) => {
     // CERES_UNIQUE_ID gets automatically assigned to children
-    var isMaster = typeof process.env.CERES_UNIQUE_ID !== 'string';
+    const isMaster = typeof process.env.CERES_UNIQUE_ID !== 'string';
 
     if (isMaster) {
       // Master
 
       // Ensure we always have an array
-      var ports = ceres.config.port instanceof Array ? ceres.config.port : [ceres.config.port];
+      const ports = ceres.config.port instanceof Array ? ceres.config.port : [ceres.config.port];
 
-      ceres.log._ceres.debug('Master forking %d instances - %s', ports.length, ports.join(', '));
-      for (var i = 0; i < ports.length; i++) {
+      ceres.log.internal.debug('Master forking %d instances - %s', ports.length, ports.join(', '));
+      for (let i = 0; i < ports.length; i += 1) {
         spawn(ceres, ports[i], i);
       }
 
       // Clean up any workers
-      ['SIGTERM', 'SIGINT'].forEach(function(signal){
-        process.on(signal, function(){
-          ceres.log._ceres.debug('Master %s received %s. Attempting to clean up workers...', process.pid, signal, {
-            pid: process.pid
-          });
-          workers.forEach(function(worker){
+      ['SIGTERM', 'SIGINT'].forEach(signal => {
+        process.on(signal, () => {
+          ceres.log.internal.debug(
+            'Master %s received %s. Attempting to clean up workers...',
+            process.pid,
+            signal,
+            {
+              pid: process.pid,
+            }
+          );
+          workers.forEach(worker => {
             worker.send({
-              restart: false
+              restart: false,
             });
             worker.kill('SIGKILL');
           });
@@ -153,7 +163,8 @@ module.exports = function(ceres) {
       resolve();
     } else {
       // Child
-      ceres.connect.call(ceres, ceres)
+      ceres.connect
+        .call(ceres, ceres)
         .then(listen.bind(ceres, ceres))
         .then(resolve)
         .catch(reject);
