@@ -1,6 +1,7 @@
-var fs = require('fs');
-var path = require('path');
-var sharedsession = require('express-socket.io-session');
+const fs = require('fs');
+const path = require('path');
+// eslint-disable-next-line import/no-extraneous-dependencies
+const sharedsession = require('express-socket.io-session');
 
 /**
  * Checks to see if sockets are configured and then sets up socket.io
@@ -9,26 +10,26 @@ var sharedsession = require('express-socket.io-session');
  * @param  {HTTP}    server
  * @return {socket.io}
  */
-module.exports = function(ceres, app, server) {
-  var clientSocketEntryPath = path.resolve(ceres.config.folders.sockets, './index.js');
+module.exports = function sockets(ceres, app, server) {
+  const clientSocketEntryPath = path.resolve(ceres.config.folders.sockets, './index.js');
 
   try {
     // Can we access the file?
     fs.accessSync(clientSocketEntryPath);
-  } catch(err) {
+  } catch (err) {
     if (err.code === 'ENOENT') {
-      ceres.log._ceres.silly('Websockets configuration not found at %s', clientSocketEntryPath);
-      return;
-    } else {
-      throw err;
+      ceres.log.internal.silly('Websockets configuration not found at %s', clientSocketEntryPath);
+      return {};
     }
+    throw err;
   }
 
   // Setup socket.io
-  var io = require('socket.io')(server);
+  // eslint-disable-next-line import/no-extraneous-dependencies
+  const io = require('socket.io')(server);
 
   // Get socket router
-  var handler = require(clientSocketEntryPath);
+  const handler = require(clientSocketEntryPath);
 
   // Connect it to client connection event
   io.on('connection', handler.bind(ceres, io, ceres.Database));
@@ -36,7 +37,7 @@ module.exports = function(ceres, app, server) {
   // Share express sessions
   io.use(sharedsession(app.get('sharedSession')));
 
-  ceres.log._ceres.silly('Websockets configured');
+  ceres.log.internal.silly('Websockets configured');
 
   return io;
 };
